Hypothesis: Household Total Income is higher if Household Head is Female
This is an exercise on pandas, graphs and a little statistics.  The main goal is to get an insight on the difference on Filipino Household Total Income based on who is the head of the family - male or female. 

Dataset used is in kaggle -  https://www.kaggle.com/grosvenpaul/family-income-and-expenditure